requirements.txt
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
pytz==2023.3

